Fix addpattern incorrectly catching all MatchErrors
This change introduces a context stack with MatchError subclasses
that pattern-matching functions must raise instead of plain MatchError.

This works the following way:

* Every time a function is decorated with @addpattern, a subclass of
MatchError is created and stored in the local scope;
* when such function is called, this subclass is put on the stack
before execution of addpattern's base_func;
* when a pattern-matching function is called, it checks whether
the exception class at the top of the stack has been already taken;
* if it wasn't, the function marks it as taken (so that no other
function would take it from the stack) and will raise it instead of
plain MatchError;
* inside addpattern, this specific exception is caught instead of
MatchError.

As a result, any pattern-matching function called directly as
addpattern's base_func will raise a specific subclass of MatchError
but will not propagate it to any pattern-matching functions called inside.

The only case when addpattern will still catch a MatchError incorrectly
is when addpattern wraps a non-pattern-matching function which calls a
pattern-matching function internally (because the non-pattern-matching
function will not mark the exception on the context stack as taken),
but this should be extremely rare because addpattern is meant only for
pattern-matching functions.

This is an alternative implementation of PR #435.